Use normal depfile rules for Rust targets
The `deps = gcc` rules delete the depfile after it has been read into
Ninja's database. This conflicts with the fact that `rustc` puts a
target for the depfile itself within the depfile, and so the deleted
depfile triggers a rebuild of the whole target.

Ninja's normal depfile rules suffice, but won't perform as well for
large Rust targets. The alternative seems to be modifying the depfile,
though, which seems more gross.